move HandleDeepLinkRequests trait to email package
